# api.yaml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: Git md diff
servers:
  - url: localhost:5000
    description: Local dev server
security:
  - JWTAuth: []
paths:
  /users/{id}:
    description: Returns the user matching the requested ID
    get:
      summary: Request info for a specific user
      tags:
        - User info
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the user to retrieve
          schema:
            type: integer
            example: 521
      responses:
        "200":
          description: Successfully retrieved data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response<UserInfo>"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response<Bool>"
              example:
                success: false
        "403":
          description: You are not authorized to perform this action
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response<Bool>"
              example:
                success: false
                
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response<Bool>"
              example:
                success: false
                message: Internal server error
...
  /auth/logout/:
    get:
      summary: Logout the user identified by token
      tags:
        - Auth
      responses:
        "200":
          description: Successfully logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response<Bool>"
              example:
                success: true

        "403":
          description: You are already logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response<Bool>"
              example:
                success: false
                
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response<Bool>"
              example:
                success: false
                message: Internal server error
  /documentations:
    get:
      summary: Returns a list of all documentations that the user has access to
      tags:
        - Documentation management

      responses:
        "200":
          description: Successfully retrieved data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response<DocumentationInfo[]>"

        "403":
          description: You are not authorized to perform this action
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response<Bool>"
              example:
                success: false
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response<Error>"
              example:
                success: false
                error: "Could not connect to the database"
    put:
      summary: Creates, updates or imports a documentation
      tags:
        - Documentation management
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentationInfo"

      responses:
        "200":
          description: Successfully created / updated documentation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response<DocumentationInfo>"

        "403":
          description: You are not authorized to perform this action
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response<Bool>"
              example:
                success: false

        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response<Error>"
              example:
                success: false
                error: "The name is already taken"
  /documentations/{id}:
    get:
      summary: Returns a documentation identified by the ID specified
      tags:
        - Documentation management
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the documentation
          schema:
            type: integer
            example: 521

      responses:
        "200":
          description: Successfully retrieved data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response<DocumentationInfo>"

        "403":
          description: You are not authorized to perform this action
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response<Bool>"
              example:
                success: false
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response<Error>"
              example:
                success: false
                
    delete:
      summary: Deletes a documentation
      tags:
        - Documentation management
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the documentation
          schema:
            type: integer
            example: 521
      responses:
        "200":
          description: Successfully deleted documentation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response<UserInfo>"

        "403":
          description: You are not authorized to perform this action
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response<Bool>"
              example:
                success: false
                
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response<Error>"
              example:
                success: false
  /documentations/{id}/versions:
    get:
      summary: Returns a list of all versions the documentation has
      tags:
        - Documentation content management
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the documentation
          schema:
            type: integer
            example: 521
      responses:
        "200":
          description: Successfully retrieved data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response<VersionInfo[]>"

        "403":
          description: You are not authorized to perform this action
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response<Bool>"
              example:
                success: false
                
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response<Error>"
              example:
                success: false
  /documentations/{id}/users:
    get:
      summary: Retrieves users from documentation
      tags:
        - Documentation management
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the documentation
          schema:
            type: integer
            example: 521
              
      responses:
        "200":
          description: Successfully retrieved data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response<UserInfo[]>"

        "403":
          description: You are not authorized to perform this action
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response<Bool>"
              example:
                success: false
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response<Error>"
              example:
                success: false
                
    put:
      summary: Adds user to a documentation
      tags:
        - Documentation management
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the documentation
          schema:
            type: integer
            example: 521
            
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleAdd"
      responses:
        "200":
          description: Successfully added user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response<UserInfo>"

        "403":
          description: You are not authorized to perform this action
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response<Bool>"
              example:
                success: false
                
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response<Error>"
              example:
                success: false
  /documentations/{id}/users/{uid}:
    delete:
      summary: Delete user from documentation
      tags:
        - Documentation management
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the documentation
          schema:
            type: integer
            example: 521
        - name: uid
          in: path
          required: true
          description: The id of the user
          schema:
            type: integer
            example: 521

      responses:
        "200":
          description: Successfully deleted user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response<UserInfo[]>"

        "403":
          description: You are not authorized to perform this action
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response<Bool>"
              example:
                success: false
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response<Error>"
              example:
                success: false
                
    put:
      summary: Edit user role in a documentation
      tags:
        - Documentation management
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the documentation
          schema:
            type: integer
            example: 521
        - name: uid
          in: path
          required: true
          description: The id of the user
          schema:
            type: integer
            example: 521
            
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleAdd"
            example:
               level: 1
               
      responses:
        "200":
          description: Successfully edited user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response<UserInfo>"

        "403":
          description: You are not authorized to perform this action
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response<Bool>"
              example:
                success: false
                
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response<Error>"
              example:
                success: false
  /documentations/provider/{provider}:
    get:
      summary: Returns data about existing repositories not present in the app, useful for import
      tags:
        - Documentation management
      parameters:
        - name: provider
          in: path
          required: true
          description: The provider slug from which to load the documentation information
          schema:
            type: string
            example: "gitlab"
              
      responses:
        "200":
          description: Successfully retrieved data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response<RemoteDocumentationInfo[]>"

        "403":
          description: You are not authorized to perform this action
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response<Bool>"
              example:
                success: false
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response<Error>"
              example:
                success: false
  /documentations/provider/{provider}/users/{search}:
    get:
      summary: Search remote users matching a specific search string
      tags:
        - Documentation management
      parameters:
        - name: provider
          in: path
          required: true
          description: The provider slug from which to load the information
          schema:
            type: string
            example: "gitlab"
        - name: search
          in: path
          required: true
          description: The search query
          schema:
            type: string
            example: "sajdlvoj"
            
      responses:
        "200":
          description: Successfully retrieved data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response<RemoteUser[]>"

        "403":
          description: You are not authorized to perform this action
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response<Bool>"
              example:
                success: false
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response<Error>"
              example:
                success: false
...
components:
  securitySchemes:
    JWTAuth:            # arbitrary name for the security scheme
      type: http
      scheme: JWT
      bearerFormat: JWT
  schemas:
...
    DocumentationInfo:
      type: object
      properties:
        id:
          type: integer
          description: "Documentation ID"
          example: 521
        providerId:
          type: string
          description: "The provider ID"
          example: "522"
        provider:
          type: string
          description: "The providers slug"
          example: "gitlab"
        name:
          type: string
          description: "Documentation display name"
          example: "Best Documentation Ever"
        slug:
          description: "Documentation URL slug"
          example: "best-docu-ever"
          type: string
        description:
          type: string
          description: "Documentation description"
          example: "This is the best documentation ever created..."
...
    Response<UserInfo>:
      type: object
      required:
        - success
      properties:
        success:
          description: "Indicates whether request was successfull"
          type: boolean
          example: true
        data:
          description: "Retreived user"
          $ref: "#/components/schemas/UserInfo"
...
    Response<DocumentationInfo>:
      type: object
      required:
        - success
      properties:
        success:
          description: "Indicates whether request was successfull"
          type: boolean
          example: true
        data:
          description: "Retreived user"
          $ref: "#/components/schemas/DocumentationInfo"
    Response<DocumentationInfo[]>:
      type: object
      required:
        - success
      properties:
        success:
          description: "Indicates whether request was successfull"
          type: boolean
          example: true
        data:
          type: array
          description: "Retreived documentation info"
          items:
            $ref: "#/components/schemas/DocumentationInfo"
...
    Response<Bool>:
      type: object
      required:
        - success
      properties:
        success:
          description: "Whether request was successfull"
          type: boolean
          example: true
    Response<Error>:
      type: object
      required:
        - success
      properties:
        success:
          description: "Whether request was successfull"
          type: boolean
          example: false
        error:
          description: "Error message"
          type: string
          example: false